refactor: Payer packages per context
