+++ grails-app/domain/com/mini/asaas/payer/Payer.groovy
@@ -1,4 +1,4 @@
-package com.mini.asaas
+package com.mini.asaas.payer
 
 import com.mini.asaas.utils.PersonEntity
 import com.mini.asaas.Customer
