+++ grails-app/domain/com/mini/asaas/payment/Payment.groovy
@@ -1,6 +1,6 @@
 package com.mini.asaas.payment
 
-import com.mini.asaas.Payer
+import com.mini.asaas.payer.Payer
 import com.mini.asaas.Customer
 import com.mini.asaas.enums.payment.BillingType
 import com.mini.asaas.enums.payment.PaymentStatus
