+++ grails-app/services/com/mini/asaas/payer/PayerService.groovy
@@ -1,10 +1,10 @@
-package com.mini.asaas
+package com.mini.asaas.payer
 
-import com.mini.asaas.Payer
+import com.mini.asaas.payer.Payer
 import com.mini.asaas.Address
 import com.mini.asaas.Customer
 import com.mini.asaas.AddressService
-import com.mini.asaas.dto.PayerDTO
+import com.mini.asaas.dto.payer.PayerDTO
 import com.mini.asaas.utils.CpfCnpjUtils
 
 import javax.transaction.Transactional
