+++ grails-app/views/payment/index.gsp
@@ -1,5 +1,5 @@
 <%@ page 
-import="com.mini.asaas.Payer"
+import="com.mini.asaas.payer.Payer"
 import="com.mini.asaas.enums.payment.BillingType"
 contentType="text/html;charset=UTF-8" %>
 <html>
